+++ src/ml/GenerateSchemaTypes.ml
@@ -39,20 +39,20 @@ type authorizationConfig = {
   baselinePath: string option;
 }
 
-type authorizationScope = AuthorizationField | AuthorizationFields
-
 type authorizationFunctionReference = {
   path: string list;
   loc: Location.t;
   fileUri: Uri.t;
-  scope: authorizationScope;
 }
 
 type publicAuthorization = {reason: string; loc: Location.t; fileUri: Uri.t}
 
+type ancestorAuthorization = {reason: string; loc: Location.t; fileUri: Uri.t}
+
 type declaredAuthorization = {
   functions: authorizationFunctionReference list;
   public: publicAuthorization option;
+  byAncestor: ancestorAuthorization option;
 }
 
 type authorizationInjection = AuthorizationContext | AuthorizationInfo
@@ -79,17 +79,20 @@ type authorizationGapKind =
 
 type effectiveAuthorizationPlan = {
   functions: authorizationFunction list;
-  inheritedScopePolicies: inheritedScopePolicy list;
+  byAncestor: ancestorAuthorization option;
+  ancestorBoundaries: ancestorAuthorizationBoundary list;
   public: publicAuthorization option;
   resolverOutcome: resolverOutcome option;
   synthetic: bool;
   baselineGap: authorizationGapKind option;
 }
 
-and inheritedScopePolicy = {
-  boundaryCoordinate: string;
-  fn: authorizationFunction;
-}
+and ancestorAuthorizationBoundary =
+  | AncestorPolicy of {boundaryCoordinate: string; fn: authorizationFunction}
+  | AncestorResolverOutcome of {
+      boundaryCoordinate: string;
+      outcome: resolverOutcome;
+    }
 
 type typeLocationLoc = {
   fileName: string;